Move websockets.hpp out of include/deepstream
