+++ src/error_handler.cpp
@@ -24,7 +24,7 @@
 #include <deepstream/error_handler.hpp>
 #include <deepstream/message.hpp>
 #include <deepstream/parser.hpp>
-#include <deepstream/websockets.hpp>
+#include "websockets.hpp"
 
 namespace deepstream {
