+++ src/websockets/poco.cpp
@@ -32,7 +32,7 @@
 #include <deepstream/buffer.hpp>
 #include <deepstream/exception.hpp>
 #include <deepstream/time.hpp>
-#include <deepstream/websockets.hpp>
+#include "../websockets.hpp"
 #include <deepstream/websockets/poco.hpp>
 
 #include <cassert>
